RAG-Chatbot/requirements.txt
langchain 
langchain-community 
langchainhub 
streamlit 
faiss-cpu 
groq 
python-dotenv
langchain-groq

